+++ nuxeo-core/nuxeo-core-bulk/src/main/java/org/nuxeo/ecm/core/bulk/action/computation/ZipBlob.java
@@ -24,11 +24,13 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.nuxeo.ecm.core.api.Blob;
 import org.nuxeo.ecm.core.bulk.BulkCodecs;
+import org.nuxeo.ecm.core.bulk.BulkService;
 import org.nuxeo.ecm.core.bulk.message.DataBucket;
 import org.nuxeo.ecm.core.utils.BlobUtils;
 import org.nuxeo.lib.stream.codec.Codec;
 import org.nuxeo.lib.stream.computation.ComputationContext;
 import org.nuxeo.lib.stream.computation.Record;
+import org.nuxeo.runtime.api.Framework;
 
 /**
  * @since 10.3
@@ -48,13 +50,14 @@ public class ZipBlob extends AbstractTransientBlobComputation {
         Codec<DataBucket> codec = BulkCodecs.getDataBucketCodec();
         DataBucket in = codec.decode(record.getData());
 
-        Blob blob = getBlob(in.getDataAsString());
+        String storeName = Framework.getService(BulkService.class).getStatus(in.getCommandId()).getAction();
+        Blob blob = getBlob(in.getDataAsString(), storeName);
         try {
             blob = BlobUtils.zip(blob, blob.getFilename() + ".zip");
         } catch (IOException e) {
             log.error("Unable to zip blob", e);
         }
-        storeBlob(blob, in.getCommandId());
+        storeBlob(blob, in.getCommandId(), storeName);
 
         DataBucket out = new DataBucket(in.getCommandId(), in.getCount(), getTransientStoreKey(in.getCommandId()));
         context.produceRecord(OUTPUT_1, Record.of(in.getCommandId(), codec.encode(out)));
