+++ nuxeo-core/nuxeo-core-test/src/main/java/org/nuxeo/ecm/core/test/DocumentSetRepositoryInit.java
@@ -16,7 +16,7 @@
  * Contributors:
  *     pierre
  */
-package org.nuxeo.ecm.core.bulk;
+package org.nuxeo.ecm.core.test;
 
 import java.util.Date;
 
@@ -28,7 +28,6 @@ import org.nuxeo.ecm.core.api.security.ACL;
 import org.nuxeo.ecm.core.api.security.ACP;
 import org.nuxeo.ecm.core.api.security.impl.ACLImpl;
 import org.nuxeo.ecm.core.api.security.impl.ACPImpl;
-import org.nuxeo.ecm.core.test.DefaultRepositoryInit;
 
 /**
  * @since 10.2
