+++ nuxeo-core/nuxeo-core-test/src/test/java/org/nuxeo/ecm/core/bulk/TestRemoveProxyAction.java
@@ -21,9 +21,9 @@ package org.nuxeo.ecm.core.bulk;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
-import static org.nuxeo.ecm.core.bulk.DocumentSetRepositoryInit.created_non_proxy;
-import static org.nuxeo.ecm.core.bulk.DocumentSetRepositoryInit.created_proxy;
-import static org.nuxeo.ecm.core.bulk.DocumentSetRepositoryInit.created_total;
+import static org.nuxeo.ecm.core.test.DocumentSetRepositoryInit.created_non_proxy;
+import static org.nuxeo.ecm.core.test.DocumentSetRepositoryInit.created_proxy;
+import static org.nuxeo.ecm.core.test.DocumentSetRepositoryInit.created_total;
 import static org.nuxeo.ecm.core.bulk.action.RemoveProxyAction.ACTION_NAME;
 
 import java.time.Duration;
@@ -38,6 +38,7 @@ import org.nuxeo.ecm.core.api.PathRef;
 import org.nuxeo.ecm.core.bulk.message.BulkCommand;
 import org.nuxeo.ecm.core.bulk.message.BulkStatus;
 import org.nuxeo.ecm.core.test.CoreFeature;
+import org.nuxeo.ecm.core.test.DocumentSetRepositoryInit;
 import org.nuxeo.ecm.core.test.annotations.RepositoryConfig;
 import org.nuxeo.runtime.test.runner.Deploy;
 import org.nuxeo.runtime.test.runner.Features;
