+++ nuxeo-distribution/nuxeo-nxr-server/pom.xml
@@ -397,6 +397,10 @@
     </dependency>
     <dependency>
       <groupId>org.nuxeo.ecm.platform</groupId>
+      <artifactId>nuxeo-platform-csv-export</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.nuxeo.ecm.platform</groupId>
       <artifactId>nuxeo-platform-forms-layout-api</artifactId>
     </dependency>
     <dependency>
