+++ nuxeo-features/nuxeo-platform-csv-export/src/main/resources/META-INF/MANIFEST.MF
@@ -1,0 +1,6 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 1
+Bundle-Name: Nuxeo Core Bulk
+Bundle-SymbolicName: org.nuxeo.ecm.platform.csv.export;singleton:=true
+Bundle-Version: 1.0.0
+Nuxeo-Component: OSGI-INF/csv-export-config.xml
