+++ nuxeo-features/nuxeo-platform-csv-export/src/main/resources/OSGI-INF/csv-export-config.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0"?>
+<component name="org.nuxeo.ecm.platform.csv.export.config" version="1.0.0">
+
+  <require>org.nuxeo.ecm.core.bulk</require>
+
+  <extension target="org.nuxeo.ecm.core.bulk" point="actions">
+    <action name="csvExport" bucketSize="100" batchSize="50"/>
+  </extension>
+
+  <extension target="org.nuxeo.runtime.stream.service" point="streamProcessor">
+    <!-- CSV exporter processor -->
+    <streamProcessor name="csvExport" class="org.nuxeo.ecm.platform.csv.export.action.CSVExportAction" logConfig="bulk"
+      defaultConcurrency="2" defaultPartitions="2">
+      <option name="produceImmediate">false</option>
+    </streamProcessor>
+  </extension>
+
+</component>
